feat: move discord bot as service
